Revert "Home view"
