+++ FoodsClassificator/FoodsClassificator/App/FoodsClassificatorApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct FoodsClassificatorApp: App {
     var body: some Scene {
         WindowGroup {
-            HomeView().preferredColorScheme(.light)
+            ContentView().preferredColorScheme(.light)
         }
     }
 }
