+++ FoodsClassificator/FoodsClassificator/Assets.xcassets/fireSymbolHome.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "Vector2.pdf",
+      "filename" : "Vector.pdf",
       "idiom" : "universal",
       "scale" : "1x"
     },
