+++ FoodsClassificator/FoodsClassificator/Assets.xcassets/fireSymbolHome.imageset/Vector.pdf
@@ -11,7 +11,7 @@ stream
 /DeviceRGB cs
 q
 1.000000 0.000000 -0.000000 1.000000 0.000000 -0.009277 cm
-1.000000 1.000000 1.000000 scn
+0.303223 0.090967 0.000000 scn
 9.485947 5.263767 m
 9.317690 4.714554 9.022785 4.215169 8.626295 3.808056 c
 8.229805 3.400942 7.743451 3.098135 7.208569 2.925370 c
@@ -77,7 +77,7 @@ endobj
 4 0 obj
   << /Annots []
      /Type /Page
-     /MediaBox [ 0.000000 0.000000 12.324585 15.405762 ]
+     /MediaBox [ 0.000000 0.000000 12.324646 15.405762 ]
      /Resources 1 0 R
      /Contents 2 0 R
      /Parent 5 0 R
