--- FoodsClassificator/FoodsClassificator/Utilities/Texts/HomeTexts.swift
@@ -1,16 +1,0 @@
-//
-//  HomeTexts.swift
-//  FoodsClassificator
-//
-//  Created by Victor Hugo Pacheco Araujo on 22/03/24.
-//
-
-import Foundation
-
-struct HomeTexts {
-    
-    static let restam = "Restam"
-    
-    static let calorias = "Calorias"
-    
-}
