--- FoodsClassificator/FoodsClassificator/Views/Home/Components/Header/UserPhotoAndStreakComponentHomeView.swift
@@ -1,60 +1,0 @@
-//
-//  UserPhotoAndStreakComponentHomeView.swift
-//  FoodsClassificator
-//
-//  Created by Victor Hugo Pacheco Araujo on 21/03/24.
-//
-
-import SwiftUI
-
-struct UserPhotoAndStreakComponentHomeView: View {
-    
-    @State private var viewModel = HomeViewModel()
-    
-    var body: some View {
-        
-        ZStack{
-            
-            HStack{
-                Text("\(viewModel.streakNumber)")
-                    .font(.callout)
-                    .bold()
-                    .foregroundStyle(Color.white)
-                Image("fireSymbolHome")
-                
-                Spacer()
-            }
-            .frame(width: 80, height: 29)
-            .padding(.horizontal)
-            .background(Color.clear)
-            .overlay(
-                RoundedRectangle(cornerRadius: 10)
-                    .stroke(Color.laranjaMacros, lineWidth: 1)
-            )
-
-            Circle()
-                .overlay {
-                    Image(viewModel.userPhoto)
-                        .resizable()
-                        .scaledToFit()
-                        .clipShape(Circle()
-                        )
-        
-                }
-                .frame(width: 45, height: 45)
-                .overlay {
-                    Circle()
-                        .stroke(lineWidth: 2)
-                        .foregroundStyle(.white)
-                }
-                .offset(x: 35)
-            
-            
-        }
-        
-    }
-}
-
-#Preview {
-    UserPhotoAndStreakComponentHomeView()
-}
