+++ FoodsClassificator/FoodsClassificator/Views/Home/Components/UserNameComponentHomeView.swift
@@ -14,7 +14,6 @@ struct UserNameComponentHomeView: View {
     var body: some View {
         Text(viewModel.userName)
             .font(.title)
-            .foregroundStyle(.white)
     }
 }
